Reorganize core tests by implementation file
